writing all matrices to individual files

+++ GZ-2D-LZ/GZ-2D-LZ.UnitTests/Gz2dlzDecoderTests.cs
@@ -73,8 +73,8 @@ namespace GZ_2D_LZ.UnitTests
             {
                 for (int x = 0; x < width; x++)
                 {
-                    Assert.AreEqual(0, matchLocation[y, x].Y);
-                    Assert.AreEqual(0, matchLocation[y, x].X);
+                    Assert.AreEqual((uint) 0, matchLocation[y, x].Y);
+                    Assert.AreEqual((uint) 0, matchLocation[y, x].X);
                 }
             }
         }
@@ -97,8 +97,8 @@ namespace GZ_2D_LZ.UnitTests
             {
                 for (int x = 0; x < width; x++)
                 {
-                    Assert.AreEqual(0, matchDimensions[y, x].Width);
-                    Assert.AreEqual(0, matchDimensions[y, x].Height);
+                    Assert.AreEqual((uint) 0, matchDimensions[y, x].Width);
+                    Assert.AreEqual((uint) 0, matchDimensions[y, x].Height);
                 }
             }
         }
@@ -203,15 +203,15 @@ namespace GZ_2D_LZ.UnitTests
 
             Assert.AreEqual(10, height);
             Assert.AreEqual(10, width);
-            Assert.AreEqual(1, matchLocation[6, 5].X);
-            Assert.AreEqual(2, matchLocation[6, 5].Y);
+            Assert.AreEqual((uint) 1, matchLocation[6, 5].X);
+            Assert.AreEqual((uint) 2, matchLocation[6, 5].Y);
 
             for (int y = 0; y < 6; y++)
             {
                 for (int x = 0; x < 5; x++)
                 {
-                    Assert.AreEqual(0, matchLocation[y, x].Y);
-                    Assert.AreEqual(0, matchLocation[y, x].X);
+                    Assert.AreEqual((uint) 0, matchLocation[y, x].Y);
+                    Assert.AreEqual((uint) 0, matchLocation[y, x].X);
                 }
             }
 
@@ -219,8 +219,8 @@ namespace GZ_2D_LZ.UnitTests
             {
                 for (int x = 6; x < width; x++)
                 {
-                    Assert.AreEqual(0, matchLocation[y, x].Y);
-                    Assert.AreEqual(0, matchLocation[y, x].X);
+                    Assert.AreEqual((uint) 0, matchLocation[y, x].Y);
+                    Assert.AreEqual((uint) 0, matchLocation[y, x].X);
                 }
             }
         }
@@ -252,15 +252,15 @@ namespace GZ_2D_LZ.UnitTests
             Assert.AreEqual(10, height);
             Assert.AreEqual(10, width);
 
-            Assert.AreEqual(4, matchDimensions[6, 5].Width);
-            Assert.AreEqual(3, matchDimensions[6, 5].Height);
+            Assert.AreEqual((uint) 4, matchDimensions[6, 5].Width);
+            Assert.AreEqual((uint) 3, matchDimensions[6, 5].Height);
 
             for (int y = 0; y < 6; y++)
             {
                 for (int x = 0; x < 5; x++)
                 {
-                    Assert.AreEqual(0, matchDimensions[y, x].Width);
-                    Assert.AreEqual(0, matchDimensions[y, x].Height);
+                    Assert.AreEqual((uint) 0, matchDimensions[y, x].Width);
+                    Assert.AreEqual((uint) 0, matchDimensions[y, x].Height);
                 }
             }
 
@@ -268,8 +268,8 @@ namespace GZ_2D_LZ.UnitTests
             {
                 for (int x = 6; x < width; x++)
                 {
-                    Assert.AreEqual(0, matchDimensions[y, x].Width);
-                    Assert.AreEqual(0, matchDimensions[y, x].Height);
+                    Assert.AreEqual((uint) 0, matchDimensions[y, x].Width);
+                    Assert.AreEqual((uint) 0, matchDimensions[y, x].Height);
                 }
             }
         }
